requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
flake8==7.1.1
isort==5.13.2
mccabe==0.7.0
numpy==2.2.0
pandas==2.2.3
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
sqlparse==0.5.3
tabula==1.0.5
typing_extensions==4.12.2
tzdata==2024.2
